[REFACTOR] - decompose processor.go into separate modules;
